homepage header/footer/searchbar styled w/ flexbox
According to instructions as per README file. I
had forgotten about earlier lessons and combining selectors
for more efficient code so I corrected this.

I also went for a simple 'center' text-align on the
content div which in hindsight was a little too simple.
Using flex worked much better as per the solution